pyproject.toml
[project]
name = "lf-nodes"
description = "Custom nodes with a touch of extra UX ✨ Client-side LLM chat, mid-workflow image editing, history for primitives, JSON manipulation, analytics nodes, CivitAI metadata fetching and saving... and more!"
version = "0.16.0"
license = {file = "LICENSE"}
dependencies = ["numpy", "onnxruntime", "openai", "opencv-python", "pandas", "piexif", "Pillow", "requests", "svgwrite", "timm", "torch", "torchvision", "tqdm", "transformers"]

[project.urls]
Repository = "https://github.com/lucafoscili/lf-nodes"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "lucafoscili"
DisplayName = "LF Nodes"
Icon = "https://github.com/lucafoscili/lucafoscili/blob/996cb0c00612e3278798c32eea80f1332277132e/lf-nodes/logo/LFN.png"
